requirements.txt
TikTokApi
python-telegram-bot
playwright